+++ enchant/index.ts
@@ -16,7 +16,7 @@ import { heroCrownEnchantMap } from './hero-crown';
 import { ignisNecklaceEnchantMap } from './ignis-necklace';
 import { rubyEnchantMap } from './ruby';
 import { bigRubyEnchantMap } from './ruby-big';
-import { runeEnchantMap } from './rune';
+import { runeEnchantMap } from './rune-aden';
 import { runeLilithEnchantMap } from './rune-lilith';
 import { talismanOfAdenEnchantMap } from './talisman-of-aden';
 import {
